Implemented no duplicate when adding items and tests, atomic length.
Implemented no duplicate items based on id (A update would be creating the id based on the data hash, as well as currently within some languages add item method getting the list of item could be done using pipeline)
Implemented tests for each one of the languages for the no duplicate items.
Fixed some issues related to run bash file.
Within Job spawner modified the way shared jobs work so its not within the main list as well as fixed the way the host and port is being set.
Python test fixed the way the host and port is being set.